Completed task and fixed bugs
Implemented FlashcardDeckCreator UserControl and ViewModel
Fixed incorrect behavior, where even though more that 1 card were selected, FlashcardDeckCreator created a deck only with the last selected card out of the list. Redid bindings on FlashcardDeckCreator to better fit MVVM and to fix binding bugs,.
Removed button to go to flashcard view from other views as it now requires a Deck to be passed into it to render.